+++ src/app/app.component.html
@@ -396,10 +396,6 @@
         <mat-icon>people_outline</mat-icon>
         <span>Brokers</span>
   </button>
-  <button mat-menu-item [routerLink]="'/listing-details'">
-        <mat-icon>videocam</mat-icon>
-        <span>Let's talk</span>
-  </button>
   <button mat-menu-item (click)="logout()">
         <mat-icon>exit_to_app</mat-icon>
         <span>Logout</span>
